+++ hardeneks/report.py
@@ -50,7 +50,18 @@ def print_instance_public_table(instances, message, docs):
             str(instance["Instances"][0]["PublicDnsName"]),
         )
 
-    console.print(Panel(table, title=message))
+    console.print(Panel(table, title=message, subtitle=docs))
+    console.print()
+
+
+def print_node_table(nodes, message, docs):
+    table = Table()
+
+    table.add_column("NodeName", style="cyan")
+
+    for node in nodes:
+        table.add_row(node.metadata.name)
+    console.print(Panel(table, title=message, subtitle=docs))
     console.print()
 
 
@@ -165,6 +176,17 @@ def print_persistent_volume_table(persistent_volumes, message, docs):
 
     for persistent_volume in persistent_volumes:
         table.add_row(persistent_volume.metadata.name, "false")
+
+    console.print(Panel(table, title=message, subtitle=docs))
+    console.print()
+
+
+def print_role_action_table(actions, message, docs):
+    table = Table()
+    table.add_column("Action", style="cyan")
+
+    for action in actions:
+        table.add_row(action)
 
     console.print(Panel(table, title=message, subtitle=docs))
     console.print()